Use 64bit offset for the first level index of btree, and remain 32bit index for the second level index. It's trading a small increasing on the index file size with 65536x larger total btree entry size. Concurrent index build, to make full use of multicore CPU.

+++ ftshelpers.hh
@@ -31,7 +31,7 @@ struct FtsIdxHeader
 {
   uint32_t signature; // First comes the signature, FTSX
   uint32_t formatVersion; // File format version
-  uint32_t chunksOffset; // The offset to chunks' storage
+  uint64_t chunksOffset; // The offset to chunks' storage
   uint32_t indexBtreeMaxElements; // Two fields from IndexInfo
   uint32_t indexRootOffset;
   uint32_t wordCount; // Number of unique words this dictionary has
